feat:(src!)
BREAKING_CHANGE:

This changes the default value of loadClusterFromSpec to true instead of false.

The motivation is that most people will want to just load the API's their cluster has access to.  If they can't do that, then it does fallback to load the included spec file.  This could be another investigation,  to see if we really need this fallback.  If not, it would get rid of a somewhat large file.